+++ src/components/navbar.js
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 function Navbar() {
   return (
     <nav>
-      <Link className='white' to='/'>
+      <Link className="white" to="/">
         <h3> __</h3>
       </Link>
       <h1>All conturies</h1>
